Moving LogHandler to its own module in logger.
